refactor buildConflictUpdateColumns util
